refactor: derive realtime tools from the acts registry
Add the shared ACTS table and project tool schemas, guards, narration, and guide copy from the single registry while preserving every existing tool name and argument shape.
