+++ packages/acts/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@sidecar/realtime",
+  "name": "@sidecar/acts",
   "version": "0.3.9",
   "private": true,
   "type": "module",
@@ -12,7 +12,6 @@
     "typecheck": "tsc --project tsconfig.json"
   },
   "dependencies": {
-    "@sidecar/acts": "workspace:*",
     "@sidecar/attention": "workspace:*",
     "@sidecar/guide": "workspace:*",
     "@sidecar/issues": "workspace:*",
