+++ packages/acts/src/acts.ts
@@ -1,0 +1,2007 @@
+/**
+ * The acts Luke can carry for the developer, named as Realtime tools, in one
+ * table. Family membership, the spoken tool count, and the schema list are
+ * derived from it; adding a tool is adding a row.
+ *
+ * The session trio are the same acts the panel's rows offer — the two writes,
+ * and the press that opens a session where its provider keeps it — and the
+ * issue pair are the two acts a connected tracker takes. Creating a workspace
+ * and the standing-ask pair — keeping the developer's ask to hear about a
+ * session, and letting it go — are the acts with no row yet to mirror. The
+ * last four are the same presses turned toward the app itself: a settings
+ * change, showing the panel, opening the feedback composer, and the Updates
+ * row's button.
+ *
+ * All run the same gauntlet: a call is validated against the observed roster
+ * (or the guide) before anything leaves the renderer, and the main process
+ * validates it again against what it observed before anything happens. The
+ * `validate` on each row is the renderer's half of that, never a substitute
+ * for the main process's. Luke is another way to ask, never a wider one.
+ */
+
+import { attentionRequestText, maximumAttentionRequestLength } from "@sidecar/attention";
+import {
+  APP_PANEL_TAB,
+  APP_SETTING_KIND,
+  APP_UPDATE_ACT,
+  APP_UPDATE_WAIT,
+  type AppGuideSetting,
+  type AppGuideSnapshot,
+  type AppGuideUpdate,
+  type AppPanelTab,
+  type AppUpdateAct,
+  appGuideSetting,
+  appToggleValue,
+  FEEDBACK_COMPOSER_KIND,
+  type FeedbackComposerKind,
+  isAppPanelTab,
+  isAppUpdateAct,
+  isFeedbackComposerKind,
+  isSessionListSort,
+  SESSION_LIST_SORT,
+  type SessionListSort,
+} from "@sidecar/guide";
+import {
+  ACT_RESULT_STATUS,
+  type IssueIdentity,
+  type IssueTransition,
+  issueCommentText,
+  type TrackedIssue,
+} from "@sidecar/issues";
+import {
+  isSessionApplicationId,
+  matchesFilterSelection,
+  maximumSessionMessageLength,
+  maximumWorkspaceNameLength,
+  type NormalizedSession,
+  type ObservedWorkspaceProject,
+  PROVIDER_ID_LIST,
+  SESSION_APPLICATION_ID,
+  SESSION_LOCATION,
+  type SessionApplicationId,
+  type SessionControl,
+  type SessionIdentity,
+  sessionMessageText,
+  supportsSessionControl,
+  WORKSPACE_TASK_SUPPORT,
+  type WorkspaceAgentModels,
+  type WorkspaceAgentSelection,
+  workspaceNameText,
+  workspaceProjectSelectionId,
+} from "@sidecar/session";
+import {
+  isRecord,
+  isWireBoolean,
+  isWireString,
+  type UnparsedWireValue,
+  type WireRecord,
+  text as wireText,
+} from "@sidecar/wire";
+export interface RealtimeFunctionCall {
+  name: string;
+  argumentsJson: string;
+}
+
+/** Which process a tool call is about: a session, an issue, or Luke himself. */
+export const REALTIME_TOOL_FAMILY = {
+  SESSION: "session",
+  ISSUE: "issue",
+  APP: "app",
+} as const;
+
+export type RealtimeToolFamily = (typeof REALTIME_TOOL_FAMILY)[keyof typeof REALTIME_TOOL_FAMILY];
+
+export const ACT_VALIDATION_TARGET = {
+  SESSION_ROSTER: "session-roster",
+  ISSUE_ROSTER: "issue-roster",
+  WORKSPACE_PROJECT: "workspace-project",
+  SETTING_ID: "setting-id",
+  UPDATE_ROW: "update-row",
+  APP_GUIDE: "app-guide",
+} as const;
+
+export type ActValidationTarget =
+  (typeof ACT_VALIDATION_TARGET)[keyof typeof ACT_VALIDATION_TARGET];
+
+/** What a validated session tool call asks for, as the bridge names it. */
+export const SESSION_TOOL_KIND = {
+  MESSAGE: "message",
+  CONTROL: "control",
+  OPEN: "open",
+  NOTICE_REQUEST: "notice-request",
+  NOTICE_WITHDRAW: "notice-withdraw",
+  READ_TRANSCRIPT: "read-transcript",
+  CREATE_WORKSPACE: "create-workspace",
+  ADD_AGENT: "add-agent",
+  RENAME_WORKSPACE: "rename-workspace",
+  RENAME_SESSION: "rename-session",
+} as const;
+
+export type SessionToolKind = (typeof SESSION_TOOL_KIND)[keyof typeof SESSION_TOOL_KIND];
+
+/** What a validated issue tool call asks for, as the bridge names it. */
+export const ISSUE_TOOL_KIND = {
+  ISSUE_STATE: "issue-state",
+  ISSUE_COMMENT: "issue-comment",
+} as const;
+
+export type IssueToolKind = (typeof ISSUE_TOOL_KIND)[keyof typeof ISSUE_TOOL_KIND];
+
+/** What a validated app tool call asks for, as the app performs it. */
+export const APP_TOOL_KIND = {
+  SETTING: "setting",
+  PANEL: "panel",
+  FEEDBACK: "feedback",
+  UPDATE: "update",
+} as const;
+
+export type AppToolKind = (typeof APP_TOOL_KIND)[keyof typeof APP_TOOL_KIND];
+
+/** The two whole-list scopes of a spoken panel ask beyond the locations. */
+export const SESSION_LIST_ALL = "all";
+export const SESSION_LIST_VOICE = "voice";
+
+/**
+ * Every value a spoken narrowing may name, fixed by the build: the whole-list
+ * scopes, every agent this build knows, and every app that associates with
+ * sessions. The chips can hold no other value — a `SessionFilter` is drawn
+ * from these same sets — so this enum is the whole vocabulary rather than a
+ * convenience, and the model picks a token from it instead of echoing the
+ * developer's words for a matcher to guess at. Which of these values narrow
+ * to anything right now is the roster's question, answered by the validator.
+ */
+const SESSION_LIST_FILTER_VALUES: readonly string[] = [
+  ...new Set<string>([
+    SESSION_LIST_ALL,
+    SESSION_LOCATION.LOCAL,
+    SESSION_LOCATION.CLOUD,
+    SESSION_LIST_VOICE,
+    ...PROVIDER_ID_LIST,
+    ...Object.values(SESSION_APPLICATION_ID),
+  ]),
+];
+
+const SESSION_LIST_FILTER_DESCRIPTION =
+  `The values to narrow the session list to: ${SESSION_LIST_ALL} for every session, ` +
+  `${SESSION_LOCATION.LOCAL} or ${SESSION_LOCATION.CLOUD} for where work runs, ` +
+  `${SESSION_LIST_VOICE} for voice chats, an agent's provider_id, or an associated app's id. ` +
+  `Values combine — ${SESSION_LOCATION.LOCAL} with an agent keeps that agent's local ` +
+  `sessions — and ${SESSION_LIST_ALL} stands alone.`;
+
+/** What one validated tool call asks for, ready for the bridge that carries it. */
+type CarriedSessionActionFields =
+  | { kind: typeof SESSION_TOOL_KIND.MESSAGE; identity: SessionIdentity; text: string }
+  | { kind: typeof SESSION_TOOL_KIND.CONTROL; identity: SessionIdentity; control: SessionControl }
+  | {
+      kind: typeof SESSION_TOOL_KIND.OPEN;
+      identity: SessionIdentity;
+      /** The one app the developer named to open it in, resolved to its id. */
+      applicationId?: SessionApplicationId;
+    }
+  | { kind: typeof SESSION_TOOL_KIND.NOTICE_REQUEST; identity: SessionIdentity; request: string }
+  | { kind: typeof SESSION_TOOL_KIND.NOTICE_WITHDRAW; identity: SessionIdentity }
+  | { kind: typeof SESSION_TOOL_KIND.READ_TRANSCRIPT; identity: SessionIdentity }
+  | {
+      kind: typeof SESSION_TOOL_KIND.CREATE_WORKSPACE;
+      providerId: string;
+      providerProjectId: string;
+      providerTargetId?: string;
+      agent?: string;
+      name?: string;
+      task?: string;
+      /** The model the developer named for this one creation, resolved to ids. */
+      agentSelection?: WorkspaceAgentSelection;
+    }
+  | {
+      kind: typeof SESSION_TOOL_KIND.ADD_AGENT;
+      identity: SessionIdentity;
+      agent: string;
+      name?: string;
+      task?: string;
+      /** The model the developer named for this one agent, as its wire id. */
+      model?: string;
+      /** The effort riding that model, when the developer named both. */
+      effort?: string;
+    }
+  | {
+      kind: typeof SESSION_TOOL_KIND.RENAME_WORKSPACE;
+      identity: SessionIdentity;
+      /** The workspace's new name, exactly as the developer chose it. */
+      name: string;
+    }
+  | {
+      kind: typeof SESSION_TOOL_KIND.RENAME_SESSION;
+      identity: SessionIdentity;
+      /** The chat's new name, exactly as the developer chose it. */
+      name: string;
+    };
+
+export type CarriedSessionAction = CarriedSessionActionFields & {
+  status?: never;
+  reason?: never;
+};
+
+type ActRejection = {
+  status: typeof ACT_RESULT_STATUS.REJECTED;
+  reason: string;
+  kind?: never;
+};
+export type SessionToolAction = CarriedSessionAction | ActRejection;
+
+/** What one validated issue tool call asks for, ready for the bridge that carries it. */
+type CarriedIssueActionFields =
+  | {
+      kind: typeof ISSUE_TOOL_KIND.ISSUE_STATE;
+      identity: IssueIdentity;
+      transition: IssueTransition;
+    }
+  | { kind: typeof ISSUE_TOOL_KIND.ISSUE_COMMENT; identity: IssueIdentity; body: string };
+
+export type CarriedIssueAction = CarriedIssueActionFields & {
+  status?: never;
+  reason?: never;
+};
+export type IssueToolAction = CarriedIssueAction | ActRejection;
+
+/**
+ * What one validated app tool call asks for, ready for the app to perform.
+ * The feedback action opens the composer and nothing else: `draft` is at most
+ * the developer's own words, placed only into an empty note, and what the
+ * composer holds leaves only by its own Send button — no action here sends.
+ */
+type CarriedAppActionFields =
+  | {
+      kind: typeof APP_TOOL_KIND.SETTING;
+      setting: AppGuideSetting;
+      value: string;
+      /** The effort riding the new value, when the developer named both. */
+      effort?: string;
+    }
+  | {
+      kind: typeof APP_TOOL_KIND.PANEL;
+      tab: AppPanelTab;
+      /** The validated narrowing, combined like the chips: OR within an axis, AND across. */
+      filters?: readonly string[];
+      sort?: SessionListSort;
+      /** Words to search the list for, exactly as the developer asked them. */
+      query?: string;
+    }
+  | { kind: typeof APP_TOOL_KIND.FEEDBACK; composer: FeedbackComposerKind; draft?: string }
+  | { kind: typeof APP_TOOL_KIND.UPDATE; act: AppUpdateAct };
+
+export type CarriedAppAction = CarriedAppActionFields & {
+  status?: never;
+  reason?: never;
+};
+export type AppToolAction = CarriedAppAction | ActRejection;
+
+export type CarriedAct = CarriedSessionAction | CarriedIssueAction | CarriedAppAction;
+
+export interface ActEnvelope {
+  id: string;
+  act: CarriedAct;
+  armed: boolean;
+}
+
+export function isActEnvelope(value: UnparsedWireValue): value is ActEnvelope & WireRecord {
+  if (
+    !isRecord(value) ||
+    !isWireString(value.id) ||
+    !isWireBoolean(value.armed) ||
+    !isRecord(value.act) ||
+    !isWireString(value.act.kind)
+  ) {
+    return false;
+  }
+  return ACTS_BY_NAME.get(value.id)?.actionKind === value.act.kind;
+}
+
+const APP_ACT_KINDS: ReadonlySet<string> = new Set(Object.values(APP_TOOL_KIND));
+const ISSUE_ACT_KINDS: ReadonlySet<string> = new Set(Object.values(ISSUE_TOOL_KIND));
+
+export function isCarriedAppAction(action: CarriedAct): action is CarriedAppAction {
+  return APP_ACT_KINDS.has(action.kind);
+}
+
+export function isCarriedIssueAction(action: CarriedAct): action is CarriedIssueAction {
+  return ISSUE_ACT_KINDS.has(action.kind);
+}
+
+export function isCarriedSessionAction(action: CarriedAct): action is CarriedSessionAction {
+  return !isCarriedAppAction(action) && !isCarriedIssueAction(action);
+}
+
+export interface SessionToolContext {
+  sessions: readonly NormalizedSession[];
+  workspaceProjects: readonly ObservedWorkspaceProject[];
+  agentModels: (providerId: string) => readonly WorkspaceAgentModels[];
+  /**
+   * The developer's saved tie-breaks for a creation ask, the same ones the
+   * projects context narrates: the provider a nameless ask goes to, and each
+   * provider's chosen project. Both only ever narrow within the listed
+   * projects — a default can settle an ambiguous ask, never widen where one
+   * can land or override a provider or project the ask actually named.
+   */
+  defaultProviderId?: string;
+  defaultProjectIds?: Readonly<Partial<Record<string, string>>>;
+}
+
+export interface IssueToolContext {
+  issues: readonly TrackedIssue[];
+}
+
+export interface AppToolContext {
+  guide: AppGuideSnapshot;
+  sessions: readonly NormalizedSession[];
+}
+
+type JsonSchemaStringProperty = {
+  type: "string";
+  description?: string;
+  enum?: readonly string[];
+};
+
+type JsonSchemaObjectProperty = {
+  type: "object";
+  description?: string;
+  properties?: JsonSchemaPropertyMap;
+  required?: readonly string[];
+  additionalProperties?: boolean;
+};
+
+type JsonSchemaArrayProperty = {
+  type: "array";
+  description?: string;
+  items: JsonSchemaStringProperty;
+};
+
+type JsonSchemaProperty =
+  | JsonSchemaStringProperty
+  | JsonSchemaObjectProperty
+  | JsonSchemaArrayProperty;
+
+type JsonSchemaPropertyMap = {
+  readonly [key: string]: JsonSchemaProperty;
+};
+
+interface RealtimeToolParameters {
+  type: "object";
+  properties: JsonSchemaPropertyMap;
+  required: readonly string[];
+}
+
+export interface RealtimeToolWireDefinition {
+  type: "function";
+  name: string;
+  description: string;
+  parameters: RealtimeToolParameters;
+}
+
+type SessionToolValidate = (parsed: WireRecord, context: SessionToolContext) => SessionToolAction;
+
+type IssueToolValidate = (parsed: WireRecord, context: IssueToolContext) => IssueToolAction;
+
+type AppToolValidate = (parsed: WireRecord, context: AppToolContext) => AppToolAction;
+
+type ActActionKind = CarriedAct["kind"];
+type ActNarration = (action: CarriedAct, sessions: readonly NormalizedSession[]) => string;
+
+interface RealtimeToolSchema {
+  description: string;
+  parameters: RealtimeToolParameters;
+}
+
+type SessionToolSpec = {
+  name: string;
+  family: typeof REALTIME_TOOL_FAMILY.SESSION;
+  actionKind: ActActionKind;
+  validatedAgainst: ActValidationTarget;
+  narration: ActNarration;
+  guide: string;
+  schema: RealtimeToolSchema;
+  validate: SessionToolValidate;
+};
+
+type IssueToolSpec = {
+  name: string;
+  family: typeof REALTIME_TOOL_FAMILY.ISSUE;
+  actionKind: ActActionKind;
+  validatedAgainst: ActValidationTarget;
+  narration: ActNarration;
+  guide: string;
+  schema: RealtimeToolSchema;
+  validate: IssueToolValidate;
+};
+
+type AppToolSpec = {
+  name: string;
+  family: typeof REALTIME_TOOL_FAMILY.APP;
+  actionKind: ActActionKind;
+  validatedAgainst: ActValidationTarget;
+  narration: ActNarration;
+  guide: string;
+  schema: RealtimeToolSchema;
+  validate: AppToolValidate;
+};
+
+type RealtimeToolSpec = SessionToolSpec | IssueToolSpec | AppToolSpec;
+
+const SESSION_IDENTITY_PARAMETERS = {
+  provider_id: {
+    type: "string",
+    description: "The session provider ID.",
+  },
+  provider_session_id: {
+    type: "string",
+    description: "The session ID.",
+  },
+} as const;
+
+const ISSUE_IDENTITY_PARAMETERS = {
+  tracker_id: {
+    type: "string",
+    description: "The tracker ID.",
+  },
+  issue_id: {
+    type: "string",
+    description: "The issue ID.",
+  },
+} as const;
+
+function textArgument(record: WireRecord, key: string): string | undefined {
+  return wireText(record[key]);
+}
+
+function parseToolArguments(
+  call: RealtimeFunctionCall,
+): { ok: true; value: WireRecord } | { ok: false; reason: string } {
+  let parsed: UnparsedWireValue;
+  try {
+    // SAFETY: JSON.parse returns a runtime value; isRecord validates the object contract.
+    parsed = JSON.parse(call.argumentsJson) as UnparsedWireValue;
+  } catch {
+    return { ok: false, reason: "The tool call's arguments were not readable." };
+  }
+  if (!isRecord(parsed)) {
+    return { ok: false, reason: "The tool call's arguments were not readable." };
+  }
+  return { ok: true, value: parsed };
+}
+
+function sessionFromArguments(
+  parsed: WireRecord,
+  sessions: readonly NormalizedSession[],
+): { session: NormalizedSession; identity: SessionIdentity } | ActRejection {
+  const providerId = textArgument(parsed, "provider_id");
+  const providerSessionId = textArgument(parsed, "provider_session_id");
+  const session = sessions.find(
+    (candidate) =>
+      candidate.providerId === providerId && candidate.providerSessionId === providerSessionId,
+  );
+  if (!session) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "No observed session matches that identity.",
+    };
+  }
+  return {
+    session,
+    identity: {
+      providerId: session.providerId,
+      providerSessionId: session.providerSessionId,
+    },
+  };
+}
+
+function issueFromArguments(
+  parsed: WireRecord,
+  issues: readonly TrackedIssue[],
+): { issue: TrackedIssue; identity: IssueIdentity } | ActRejection {
+  const trackerId = textArgument(parsed, "tracker_id");
+  const issueId = textArgument(parsed, "issue_id");
+  const issue = issues.find(
+    (candidate) => candidate.trackerId === trackerId && candidate.identifier === issueId,
+  );
+  if (!issue) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "No tracked issue matches that identity.",
+    };
+  }
+  return {
+    issue,
+    identity: {
+      trackerId: issue.trackerId,
+      identifier: issue.identifier,
+    },
+  };
+}
+
+/**
+ * Resolves a model the developer named — by the label the guide lists it
+ * under, or its id — to the wire pairing an endpoint takes, held to the
+ * build's documented entries for the provider. The effort, when named, must
+ * be one the resolved model's own agent documents: the pairing is validated
+ * as the whole it will be sent as.
+ */
+function resolveWorkspaceAgentModel(
+  entries: readonly WorkspaceAgentModels[],
+  modelWord: string,
+  effortWord: string | undefined,
+): { selection: WorkspaceAgentSelection } | { refusal: string } {
+  const normalizedModel = modelWord.trim().toLowerCase();
+  const named = entries
+    .flatMap((entry) => entry.models.map((model) => ({ entry, model })))
+    .find(
+      ({ model }) =>
+        model.label.toLowerCase() === normalizedModel || model.id.toLowerCase() === normalizedModel,
+    );
+  if (!named) return { refusal: "No documented model goes by that name here." };
+  let effort: string | undefined;
+  if (effortWord !== undefined) {
+    const normalizedEffort = effortWord.trim().toLowerCase();
+    effort = named.entry.efforts.find((candidate) => candidate.toLowerCase() === normalizedEffort);
+    if (!effort) {
+      return {
+        refusal:
+          named.entry.efforts.length > 0
+            ? `That model's effort is one of ${named.entry.efforts.join(", ")}.`
+            : "That model takes no effort level.",
+      };
+    }
+  }
+  const selection: WorkspaceAgentSelection = {
+    agent: named.entry.agent,
+    model: named.model.id,
+  };
+  if (effort) selection.effort = effort;
+  return { selection };
+}
+
+function validateSendSessionMessage(
+  parsed: WireRecord,
+  context: SessionToolContext,
+): SessionToolAction {
+  const found = sessionFromArguments(parsed, context.sessions);
+  if ("status" in found) return found;
+  const { session, identity } = found;
+  if (!session.canReceiveMessage) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "That session does not take messages right now.",
+    };
+  }
+  const messageText = sessionMessageText(parsed.text);
+  if (!messageText) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "That message is empty or too long.",
+    };
+  }
+  return { kind: SESSION_TOOL_KIND.MESSAGE, identity, text: messageText };
+}
+
+function validateRunSessionControl(
+  parsed: WireRecord,
+  context: SessionToolContext,
+): SessionToolAction {
+  const found = sessionFromArguments(parsed, context.sessions);
+  if ("status" in found) return found;
+  const { session, identity } = found;
+  const controlId = textArgument(parsed, "control_id");
+  const control = session.controls.find((candidate) => candidate.id === controlId);
+  if (!controlId || !control || !supportsSessionControl(session, controlId)) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "That session advertises no such control.",
+    };
+  }
+  return { kind: SESSION_TOOL_KIND.CONTROL, identity, control };
+}
+
+function validateOpenSession(parsed: WireRecord, context: SessionToolContext): SessionToolAction {
+  const found = sessionFromArguments(parsed, context.sessions);
+  if ("status" in found) return found;
+  const { session, identity } = found;
+  // The action carries the identity — and, when the developer named an app,
+  // that app's id — never the address: the main process reads the link back
+  // out of its own registry, the same as a pressed row or a pressed app mark.
+  const applicationWord = textArgument(parsed, "application");
+  if (applicationWord !== undefined) {
+    const normalized = applicationWord.trim().toLowerCase();
+    const application = session.applications.find(
+      (candidate) =>
+        candidate.displayName.toLowerCase() === normalized || candidate.id === normalized,
+    );
+    // An association without an exact address identifies the app but opens
+    // nothing, so it refuses like an app the roster never listed — and the
+    // refusal names the apps that can open, which the roster already carries.
+    // The id must be one the build fixed: the bridge takes no other, and the
+    // main process would refuse it again.
+    const applicationId = application?.link ? application.id : undefined;
+    if (applicationId === undefined || !isSessionApplicationId(applicationId)) {
+      const openable = session.applications.filter((candidate) => candidate.link);
+      return {
+        status: ACT_RESULT_STATUS.REJECTED,
+        reason:
+          openable.length > 0
+            ? `That session opens in ${openable
+                .map((candidate) => candidate.displayName)
+                .join(" or ")}, not there.`
+            : "No app carries an exact address for that session.",
+      };
+    }
+    return { kind: SESSION_TOOL_KIND.OPEN, identity, applicationId };
+  }
+  if (!session.detail.link) {
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: "That session has no address to open." };
+  }
+  return { kind: SESSION_TOOL_KIND.OPEN, identity };
+}
+
+function validateRequestSessionNotice(
+  parsed: WireRecord,
+  context: SessionToolContext,
+): SessionToolAction {
+  const found = sessionFromArguments(parsed, context.sessions);
+  if ("status" in found) return found;
+  const { identity } = found;
+  // Observation is the only prerequisite: the ask writes nothing anywhere and
+  // asks the session for nothing, so a session that takes no messages and
+  // advertises no controls can still be asked about.
+  const request = attentionRequestText(parsed.request);
+  if (!request) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: `An ask has to be under ${maximumAttentionRequestLength} characters and longer than nothing.`,
+    };
+  }
+  return { kind: SESSION_TOOL_KIND.NOTICE_REQUEST, identity, request };
+}
+
+function validateWithdrawSessionNotice(
+  parsed: WireRecord,
+  context: SessionToolContext,
+): SessionToolAction {
+  const found = sessionFromArguments(parsed, context.sessions);
+  if ("status" in found) return found;
+  return { kind: SESSION_TOOL_KIND.NOTICE_WITHDRAW, identity: found.identity };
+}
+
+function validateReadSessionTranscript(
+  parsed: WireRecord,
+  context: SessionToolContext,
+): SessionToolAction {
+  const found = sessionFromArguments(parsed, context.sessions);
+  if ("status" in found) return found;
+  const { session, identity } = found;
+  // The action carries the identity, never a path: the main process locates
+  // the transcript in its own provider home, the same way a pressed row's
+  // open never carries an address. Only a session on this machine has a
+  // transcript here to read; which local providers keep a readable one is
+  // the main process's answer.
+  if (session.location !== SESSION_LOCATION.LOCAL) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "Only local sessions keep a transcript on this machine.",
+    };
+  }
+  return { kind: SESSION_TOOL_KIND.READ_TRANSCRIPT, identity };
+}
+
+function validateCreateWorkspace(
+  parsed: WireRecord,
+  context: SessionToolContext,
+): SessionToolAction {
+  // A creation ask names a project rather than a session, so it is validated
+  // against the projects the conversation was shown — the same discipline,
+  // against the list that actually offered it.
+  const providerId = textArgument(parsed, "provider_id");
+  const projectId = textArgument(parsed, "project_id");
+  const targetId = textArgument(parsed, "target_id");
+  let matchingProjects = context.workspaceProjects.filter(
+    (candidate) =>
+      (!providerId || candidate.providerId === providerId) &&
+      (!projectId || candidate.providerProjectId === projectId) &&
+      (!targetId || candidate.providerTargetId === targetId),
+  );
+  // The saved defaults settle only what the ask left unnamed: no provider
+  // named sends a still-ambiguous ask to the default provider while it is
+  // offering, and no project named sends it on to that provider's chosen
+  // project. Neither step can leave the listed set, and an ask that named
+  // its own provider or project is never overridden.
+  if (!providerId && context.defaultProviderId && matchingProjects.length > 1) {
+    const offeredByDefault = matchingProjects.filter(
+      (candidate) => candidate.providerId === context.defaultProviderId,
+    );
+    if (offeredByDefault.length > 0) matchingProjects = offeredByDefault;
+  }
+  // A provider's chosen project settles which project, never which provider:
+  // while candidates still span providers, one provider's saved project must
+  // not quietly decide an ask the developer left open between them.
+  const [firstMatch] = matchingProjects;
+  const oneProviderMatches =
+    firstMatch !== undefined &&
+    matchingProjects.every((candidate) => candidate.providerId === firstMatch.providerId);
+  if (!projectId && oneProviderMatches && matchingProjects.length > 1) {
+    const chosenProjects = matchingProjects.filter(
+      (candidate) =>
+        context.defaultProjectIds?.[candidate.providerId] ===
+        workspaceProjectSelectionId(candidate),
+    );
+    if (chosenProjects.length === 1) matchingProjects = chosenProjects;
+  }
+  if (matchingProjects.length !== 1) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason:
+        matchingProjects.length === 0
+          ? "No listed project matches that identity."
+          : "More than one listed project matches; name the project and host.",
+    };
+  }
+  const project = matchingProjects[0];
+  if (!project)
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "No listed project matches that identity.",
+    };
+  const requestedAgent = textArgument(parsed, "agent");
+  const matchingAgents = requestedAgent
+    ? project.spawnableAgents?.filter(
+        (candidate) => candidate.toLocaleLowerCase() === requestedAgent.toLocaleLowerCase(),
+      )
+    : undefined;
+  const agent =
+    (requestedAgent && project.spawnableAgents?.includes(requestedAgent)
+      ? requestedAgent
+      : matchingAgents?.length === 1
+        ? matchingAgents[0]
+        : undefined) ?? project.defaultAgent;
+  if (project.spawnableAgents && (!agent || !project.spawnableAgents.includes(agent))) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: agent
+        ? "That project lists no such agent to start."
+        : "Name one of the agents that project lists for a new workspace.",
+    };
+  }
+  let name: string | undefined;
+  if (parsed.name !== undefined) {
+    name = workspaceNameText(parsed.name);
+    if (!name) {
+      return {
+        status: ACT_RESULT_STATUS.REJECTED,
+        reason: `A workspace name has to be under ${maximumWorkspaceNameLength} characters and longer than nothing.`,
+      };
+    }
+  }
+  // The task is held to the project's own word for it: a project that takes
+  // none cannot be handed one, a project that needs one cannot be created
+  // without it, and the text itself is bounded like the message it is.
+  let task: string | undefined;
+  if (parsed.task !== undefined) {
+    if (project.taskSupport === WORKSPACE_TASK_SUPPORT.NONE) {
+      return { status: ACT_RESULT_STATUS.REJECTED, reason: "That project takes no opening task." };
+    }
+    task = sessionMessageText(parsed.task);
+    if (!task) {
+      return {
+        status: ACT_RESULT_STATUS.REJECTED,
+        reason: "That task is empty or too long.",
+      };
+    }
+  } else if (project.taskSupport === WORKSPACE_TASK_SUPPORT.REQUIRED) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "That project needs an opening task to create a workspace.",
+    };
+  }
+  // A model named for this one creation resolves against the provider's own
+  // documented table, and the effort only ever rides a model: alone it has
+  // nothing documented to attach to.
+  const spokenModel = textArgument(parsed, "model");
+  const spokenEffort = textArgument(parsed, "effort");
+  if (spokenEffort !== undefined && spokenModel === undefined) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "An effort rides a model; name the model too.",
+    };
+  }
+  let agentSelection: WorkspaceAgentSelection | undefined;
+  if (spokenModel !== undefined) {
+    const resolved = resolveWorkspaceAgentModel(
+      context.agentModels(project.providerId),
+      spokenModel,
+      spokenEffort,
+    );
+    if ("refusal" in resolved)
+      return { status: ACT_RESULT_STATUS.REJECTED, reason: resolved.refusal };
+    agentSelection = resolved.selection;
+  }
+  const action: CarriedSessionAction = {
+    kind: SESSION_TOOL_KIND.CREATE_WORKSPACE,
+    providerId: project.providerId,
+    providerProjectId: project.providerProjectId,
+  };
+  if (project.providerTargetId) action.providerTargetId = project.providerTargetId;
+  if (agent) action.agent = agent;
+  if (name) action.name = name;
+  if (task) action.task = task;
+  if (agentSelection) action.agentSelection = agentSelection;
+  return action;
+}
+
+function validateAddWorkspaceAgent(
+  parsed: WireRecord,
+  context: SessionToolContext,
+): SessionToolAction {
+  const found = sessionFromArguments(parsed, context.sessions);
+  if ("status" in found) return found;
+  const { session, identity } = found;
+  // The agent must be one this session's own roster entry listed: the list
+  // is the provider's word for what its endpoint takes, so an ask outside it
+  // is refused rather than forwarded to be refused.
+  const agent = textArgument(parsed, "agent");
+  if (!agent || !session.spawnableAgents.includes(agent)) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "That session lists no such agent to add.",
+    };
+  }
+  let name: string | undefined;
+  if (parsed.name !== undefined) {
+    name = workspaceNameText(parsed.name);
+    if (!name) {
+      return {
+        status: ACT_RESULT_STATUS.REJECTED,
+        reason: `A session name has to be under ${maximumWorkspaceNameLength} characters and longer than nothing.`,
+      };
+    }
+  }
+  let task: string | undefined;
+  if (parsed.task !== undefined) {
+    task = sessionMessageText(parsed.task);
+    if (!task) {
+      return {
+        status: ACT_RESULT_STATUS.REJECTED,
+        reason: "That task is empty or too long.",
+      };
+    }
+  }
+  // A model named for this one agent resolves within the asked-for kind
+  // alone: the developer's chosen agent is never re-decided by the model
+  // they named beside it, so a mismatch is a refusal rather than a swap.
+  const spokenModel = textArgument(parsed, "model");
+  const spokenEffort = textArgument(parsed, "effort");
+  if (spokenEffort !== undefined && spokenModel === undefined) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "An effort rides a model; name the model too.",
+    };
+  }
+  let selection: WorkspaceAgentSelection | undefined;
+  if (spokenModel !== undefined) {
+    const entries = context
+      .agentModels(session.providerId)
+      .filter((candidate) => candidate.agent === agent);
+    const resolved = resolveWorkspaceAgentModel(entries, spokenModel, spokenEffort);
+    if ("refusal" in resolved) {
+      return {
+        status: ACT_RESULT_STATUS.REJECTED,
+        reason: resolved.refusal.startsWith("No documented model")
+          ? `A ${agent} agent runs no model by that name.`
+          : resolved.refusal,
+      };
+    }
+    selection = resolved.selection;
+  }
+  const action: CarriedSessionAction = {
+    kind: SESSION_TOOL_KIND.ADD_AGENT,
+    identity,
+    agent,
+  };
+  if (name) action.name = name;
+  if (task) action.task = task;
+  if (selection) action.model = selection.model;
+  if (selection?.effort) action.effort = selection.effort;
+  return action;
+}
+
+function validateRenameWorkspace(
+  parsed: WireRecord,
+  context: SessionToolContext,
+): SessionToolAction {
+  const found = sessionFromArguments(parsed, context.sessions);
+  if ("status" in found) return found;
+  const { session, identity } = found;
+  // Only a session whose roster entry advertised renaming has a workspace a
+  // rename can land on. The action carries the identity and the name, never
+  // the target: the main process resolves the workspace from its own
+  // registry, the same way an open never carries an address.
+  if (!session.renameTarget) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "That session's workspace cannot be renamed.",
+    };
+  }
+  const name = workspaceNameText(parsed.name);
+  if (!name) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: `A workspace name has to be under ${maximumWorkspaceNameLength} characters and longer than nothing.`,
+    };
+  }
+  return { kind: SESSION_TOOL_KIND.RENAME_WORKSPACE, identity, name };
+}
+
+function validateRenameSession(parsed: WireRecord, context: SessionToolContext): SessionToolAction {
+  const found = sessionFromArguments(parsed, context.sessions);
+  if ("status" in found) return found;
+  const { session, identity } = found;
+  if (!session.canRename) {
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: "That chat cannot be renamed." };
+  }
+  const name = workspaceNameText(parsed.name);
+  if (!name) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: `A chat name has to be under ${maximumWorkspaceNameLength} characters and longer than nothing.`,
+    };
+  }
+  return { kind: SESSION_TOOL_KIND.RENAME_SESSION, identity, name };
+}
+
+function validateUpdateIssueState(parsed: WireRecord, context: IssueToolContext): IssueToolAction {
+  const found = issueFromArguments(parsed, context.issues);
+  if ("status" in found) return found;
+  const { issue, identity } = found;
+  const state = textArgument(parsed, "state");
+  // Spoken names arrive with their case retold rather than copied, so the
+  // match forgives case alone — never spelling — and only while it stays
+  // unambiguous. Two advertised states apart only in case are not a guess
+  // Luke gets to make.
+  const named = state
+    ? issue.transitions.filter((candidate) => candidate.name.toLowerCase() === state.toLowerCase())
+    : [];
+  const transition =
+    named.find((candidate) => candidate.name === state) ??
+    (named.length === 1 ? named[0] : undefined);
+  if (!transition) {
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: "That issue lists no such state." };
+  }
+  return { kind: ISSUE_TOOL_KIND.ISSUE_STATE, identity, transition };
+}
+
+function validateCommentOnIssue(parsed: WireRecord, context: IssueToolContext): IssueToolAction {
+  const found = issueFromArguments(parsed, context.issues);
+  if ("status" in found) return found;
+  const { issue, identity } = found;
+  if (!issue.canComment) {
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: "That issue does not take comments." };
+  }
+  const body = issueCommentText(parsed.body);
+  if (!body) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "That comment is empty or too long.",
+    };
+  }
+  return { kind: ISSUE_TOOL_KIND.ISSUE_COMMENT, identity, body };
+}
+
+/**
+ * Validates the value a spoken change carries against the setting it names.
+ * A toggle takes the guide's own two words (and their unambiguous synonyms);
+ * a choice takes exactly one of the values the guide listed. Anything else is
+ * refused with the accepted set, so the refusal is also the correction.
+ */
+function appSettingValue(setting: AppGuideSetting, value: UnparsedWireValue): string | undefined {
+  if (setting.kind === APP_SETTING_KIND.TOGGLE) return appToggleValue(value);
+  if (!isWireString(value)) return undefined;
+  const normalized = value.trim().toLowerCase();
+  return setting.choices?.find((choice) => choice.toLowerCase() === normalized);
+}
+
+/**
+ * Whether one observed session answers one spoken filter value. Every
+ * identity a row carries is a filter on the same terms as its provider id —
+ * the agent behind a hosted chat, an app associated with it, and a workspace
+ * manager's scope id — so a spoken ask reaches exactly the rows the matching
+ * chip would keep.
+ */
+function sessionAnswersFilter(session: NormalizedSession, filter: string): boolean {
+  if (filter === SESSION_LOCATION.LOCAL || filter === SESSION_LOCATION.CLOUD) {
+    return session.location === filter;
+  }
+  if (filter === SESSION_LIST_VOICE) return session.realtimeVoice === true;
+  return (
+    session.providerId === filter ||
+    session.agent?.id === filter ||
+    session.workspace?.scopeId === filter ||
+    session.applications.some((application) => application.id === filter)
+  );
+}
+
+/**
+ * Validates a spoken session-list narrowing against the sessions actually
+ * being observed. A narrowing that would show nothing is refused rather than
+ * applied: the panel would quietly fall back to showing everything, and Luke
+ * would have reported a narrowing that never happened. Each value is checked
+ * on its own first, so the refusal can name the value that is wrong rather
+ * than only the combination — and then the combination is checked whole, on
+ * the same axis terms the chips combine on, because two values a roster
+ * answers separately can still name an intersection nothing occupies.
+ */
+function panelFiltersAction(
+  filters: readonly string[],
+  sessions: readonly NormalizedSession[],
+): { filters: readonly string[] } | { reason: string } {
+  // The enum on the tool's own schema already binds a compliant model to
+  // these tokens; this is the backstop for a call composed past it.
+  for (const filter of filters) {
+    if (!SESSION_LIST_FILTER_VALUES.includes(filter)) {
+      return { reason: `"${filter}" is not one of the filter values the tool lists.` };
+    }
+  }
+  const chosen = [...new Set(filters)];
+  if (chosen.includes(SESSION_LIST_ALL)) {
+    if (chosen.length > 1) {
+      return { reason: `${SESSION_LIST_ALL} is the whole list, so it combines with nothing.` };
+    }
+    return { filters: chosen };
+  }
+  for (const filter of chosen) {
+    if (sessions.some((session) => sessionAnswersFilter(session, filter))) continue;
+    if (filter === SESSION_LOCATION.LOCAL || filter === SESSION_LOCATION.CLOUD) {
+      return { reason: `No ${filter} sessions are observed right now.` };
+    }
+    if (filter === SESSION_LIST_VOICE) {
+      return { reason: "No voice sessions are observed right now." };
+    }
+    return {
+      reason: `No observed session belongs to an agent, app, or workspace manager "${filter}".`,
+    };
+  }
+  if (
+    chosen.length > 1 &&
+    !sessions.some((session) =>
+      matchesFilterSelection(chosen, (filter) => sessionAnswersFilter(session, filter)),
+    )
+  ) {
+    return { reason: "No observed session matches that combination of filters." };
+  }
+  return { filters: chosen };
+}
+
+function validateChangeAppSetting(parsed: WireRecord, context: AppToolContext): AppToolAction {
+  const setting = appGuideSetting(context.guide, textArgument(parsed, "setting_id"));
+  if (!setting) {
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: "The app guide lists no such setting." };
+  }
+  if (!setting.adjustable) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: `${setting.label} can only be changed by hand: ${setting.manual}`,
+    };
+  }
+  const value = appSettingValue(setting, parsed.value);
+  if (value === undefined) {
+    const accepted =
+      setting.kind === APP_SETTING_KIND.TOGGLE ? "on or off" : (setting.choices ?? []).join(", ");
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: `${setting.label} takes ${accepted}.` };
+  }
+  // An effort may ride only a value the guide lists levels for, so both
+  // halves of one stored pairing can be asked for in one change — matched
+  // like the value: case retold rather than copied, answered in the guide's
+  // own casing.
+  const effortWord = textArgument(parsed, "effort");
+  if (effortWord === undefined) return { kind: APP_TOOL_KIND.SETTING, setting, value };
+  const levels = setting.efforts?.[value] ?? [];
+  if (levels.length === 0) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason:
+        setting.efforts === undefined
+          ? `${setting.label} takes no effort level.`
+          : `${value} takes no effort level.`,
+    };
+  }
+  const normalizedEffort = effortWord.trim().toLowerCase();
+  const effort = levels.find((candidate) => candidate.toLowerCase() === normalizedEffort);
+  if (effort === undefined) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: `${value}'s effort is one of ${levels.join(", ")}.`,
+    };
+  }
+  return { kind: APP_TOOL_KIND.SETTING, setting, value, effort };
+}
+
+/**
+ * Reads the narrowing a panel ask carries: several values, or a lone string
+ * for a narrowing of one. Blank entries are dropped rather than validated,
+ * and an emptied list is no narrowing at all.
+ */
+function spokenFilterValues(
+  value: UnparsedWireValue,
+): { values: readonly string[] | undefined } | { reason: string } {
+  if (value === undefined) return { values: undefined };
+  const entries = isWireString(value) ? [value] : value;
+  if (!Array.isArray(entries) || !entries.every((entry) => isWireString(entry))) {
+    return { reason: "filters takes a list of filter values." };
+  }
+  const cleaned = entries.map((entry) => entry.trim()).filter((entry) => entry.length > 0);
+  return { values: cleaned.length > 0 ? cleaned : undefined };
+}
+
+function validateShowPanel(parsed: WireRecord, context: AppToolContext): AppToolAction {
+  const tab = parsed.tab ?? APP_PANEL_TAB.SESSIONS;
+  if (!isAppPanelTab(tab)) {
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: "The panel has no such tab." };
+  }
+  const sort = textArgument(parsed, "sort");
+  if (sort !== undefined && !isSessionListSort(sort)) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "The list orders by urgency or by recency.",
+    };
+  }
+  // A search is bounded by the hand's own control: the magnifier is only
+  // offered beside a list with more than one session, and a spoken ask
+  // reaches no further than it. The words themselves are not validated
+  // against the rows the way a filter is — a query is read against the lines
+  // as the surface words them, which only the renderer knows, and a search
+  // matching nothing is the list's own honest answer rather than a stale
+  // narrowing to refuse.
+  const query = textArgument(parsed, "query");
+  if (query !== undefined && context.sessions.length < 2) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "The list offers a search only when more than one session is observed.",
+    };
+  }
+  const asked = spokenFilterValues(parsed.filters);
+  if ("reason" in asked) return { status: ACT_RESULT_STATUS.REJECTED, reason: asked.reason };
+  if (asked.values === undefined) {
+    const action: CarriedAppAction = { kind: APP_TOOL_KIND.PANEL, tab };
+    if (sort !== undefined) action.sort = sort;
+    if (query !== undefined) action.query = query;
+    return action;
+  }
+  const outcome = panelFiltersAction(asked.values, context.sessions);
+  if ("reason" in outcome) return { status: ACT_RESULT_STATUS.REJECTED, reason: outcome.reason };
+  const action: CarriedAppAction = {
+    kind: APP_TOOL_KIND.PANEL,
+    tab,
+    filters: outcome.filters,
+  };
+  if (sort !== undefined) action.sort = sort;
+  if (query !== undefined) action.query = query;
+  return action;
+}
+
+function validateOpenFeedbackComposer(parsed: WireRecord, _context: AppToolContext): AppToolAction {
+  const composer = parsed.kind;
+  if (!isFeedbackComposerKind(composer)) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "The composer writes feedback or a prompt, nothing else.",
+    };
+  }
+  // The draft is the developer's ask restated in their words, not a document,
+  // so it is bounded like a typed one; a blank draft is no draft, and the
+  // composer simply opens empty.
+  const draft = textArgument(parsed, "draft")?.slice(0, maximumSessionMessageLength);
+  const action: CarriedAppAction = { kind: APP_TOOL_KIND.FEEDBACK, composer };
+  if (draft) action.draft = draft;
+  return action;
+}
+
+/**
+ * What stands where the asked-for act would be, so a refusal can say why the
+ * row is not offering it — the row's own detail says where the build stands,
+ * and this names the one press that stands instead.
+ */
+const UPDATE_BUTTON_STANDING = {
+  [APP_UPDATE_ACT.CHECK]: "Its button offers a check right now.",
+  [APP_UPDATE_ACT.DOWNLOAD]: "Its button offers the releases page in the browser right now.",
+  [APP_UPDATE_ACT.RESTART]: "Its button offers Restart to update right now.",
+  [APP_UPDATE_WAIT.CHECKING]: "Nothing is pressable while the check is out.",
+  [APP_UPDATE_WAIT.DOWNLOADING]: "Nothing is pressable while the download runs.",
+} as const satisfies Record<AppGuideUpdate["button"], string>;
+
+function validateRunUpdateAction(parsed: WireRecord, context: AppToolContext): AppToolAction {
+  // The guide's update entry is the roster here: a run that reported nothing
+  // about updates — a fixture, a pure caller — advertises no act to run.
+  const update = context.guide.update;
+  if (!update) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "This run does not report where updates stand.",
+    };
+  }
+  const act = parsed.action;
+  if (!isAppUpdateAct(act)) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: "The Updates button checks, downloads, or restarts.",
+    };
+  }
+  // One button, one act: only the press the row is actually drawing runs, so
+  // the refusal is the row's own words plus what stands in the act's place.
+  if (act !== update.button) {
+    return {
+      status: ACT_RESULT_STATUS.REJECTED,
+      reason: `${update.detail} ${UPDATE_BUTTON_STANDING[update.button]}`,
+    };
+  }
+  return { kind: APP_TOOL_KIND.UPDATE, act };
+}
+
+function observedSessionName(
+  identity: SessionIdentity,
+  sessions: readonly NormalizedSession[],
+): string {
+  const session = sessions.find(
+    (candidate) =>
+      candidate.providerId === identity.providerId &&
+      candidate.providerSessionId === identity.providerSessionId,
+  );
+  return session ? `"${session.title}"` : "a session";
+}
+
+function observedApplicationName(
+  identity: SessionIdentity,
+  applicationId: SessionApplicationId,
+  sessions: readonly NormalizedSession[],
+): string {
+  const session = sessions.find(
+    (candidate) =>
+      candidate.providerId === identity.providerId &&
+      candidate.providerSessionId === identity.providerSessionId,
+  );
+  return (
+    session?.applications.find((application) => application.id === applicationId)?.displayName ??
+    applicationId
+  );
+}
+
+function narrate<K extends ActActionKind>(
+  kind: K,
+  render: (
+    action: Extract<CarriedAct, { kind: K }>,
+    sessions: readonly NormalizedSession[],
+  ) => string,
+): ActNarration {
+  return (action, sessions) => {
+    if (action.kind !== kind) return "carried an act";
+    // SAFETY: the equality above narrows the discriminated union to the kind
+    // captured beside this renderer; TypeScript cannot correlate generic K.
+    return render(action as Extract<CarriedAct, { kind: K }>, sessions);
+  };
+}
+
+/**
+ * The acts Luke can carry, keyed the way a value set is: adding a tool is
+ * adding a key, and the family sets, the spoken count, and the schema list
+ * follow. Each row's `validate` is the renderer's half of the gauntlet — the
+ * main process still validates the same act against what it observed.
+ */
+export const ACTS = {
+  SEND_SESSION_MESSAGE: {
+    name: "send_session_message",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.MESSAGE,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(
+      SESSION_TOOL_KIND.MESSAGE,
+      (action, sessions) =>
+        `sent a message to ${observedSessionName(action.identity, sessions)}: "${action.text}"`,
+    ),
+    guide: "Send the developer's words to a session that currently advertises messages.",
+    schema: {
+      description: "Send a message to an observed session.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          text: {
+            type: "string",
+            description: "The message to send.",
+          },
+        },
+        required: ["provider_id", "provider_session_id", "text"],
+      },
+    },
+    validate: validateSendSessionMessage,
+  },
+  RUN_SESSION_CONTROL: {
+    name: "run_session_control",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.CONTROL,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(
+      SESSION_TOOL_KIND.CONTROL,
+      (action, sessions) =>
+        `ran "${action.control.label}" on ${observedSessionName(action.identity, sessions)}`,
+    ),
+    guide: "Run only a control the session currently advertises.",
+    schema: {
+      description: "Run a control advertised by an observed session.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          control_id: {
+            type: "string",
+            description: "The control ID.",
+          },
+        },
+        required: ["provider_id", "provider_session_id", "control_id"],
+      },
+    },
+    validate: validateRunSessionControl,
+  },
+  OPEN_SESSION: {
+    name: "open_session",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.OPEN,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(SESSION_TOOL_KIND.OPEN, (action, sessions) => {
+      const name = observedSessionName(action.identity, sessions);
+      return action.applicationId
+        ? `opened ${name} in ${observedApplicationName(action.identity, action.applicationId, sessions)}`
+        : `opened ${name}`;
+    }),
+    guide: "Open an observed session at the address its latest roster row reports.",
+    schema: {
+      description:
+        "Open one observed session where its provider keeps it — only when the developer asks " +
+        "to open, go to, or jump into that specific session. An ask to show, see, or list " +
+        'sessions or agents — "show me the cloud agents" — filters the panel through ' +
+        "show_panel instead, never this. An ask to open one session per provider uses this tool " +
+        "once per matching provider in the same response, without filtering the panel first.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          application: {
+            type: "string",
+            description:
+              "The app to open the session in, as its roster line's opens_in lists it — only " +
+              "when the developer named one. Omitted, the session opens at its own address.",
+          },
+        },
+        required: ["provider_id", "provider_session_id"],
+      },
+    },
+    validate: validateOpenSession,
+  },
+  REQUEST_SESSION_NOTICE: {
+    name: "request_session_notice",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.NOTICE_REQUEST,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(
+      SESSION_TOOL_KIND.NOTICE_REQUEST,
+      (action, sessions) =>
+        `remembered a standing ask about ${observedSessionName(action.identity, sessions)}: "${action.request}"`,
+    ),
+    guide: "Remember the developer's bounded standing ask for an observed session.",
+    schema: {
+      description: "Remember a request to be notified about an observed session.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          request: {
+            type: "string",
+            description: "What to notify the user about.",
+          },
+        },
+        required: ["provider_id", "provider_session_id", "request"],
+      },
+    },
+    validate: validateRequestSessionNotice,
+  },
+  WITHDRAW_SESSION_NOTICE: {
+    name: "withdraw_session_notice",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.NOTICE_WITHDRAW,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(
+      SESSION_TOOL_KIND.NOTICE_WITHDRAW,
+      (action, sessions) =>
+        `withdrew the standing ask about ${observedSessionName(action.identity, sessions)}`,
+    ),
+    guide: "Withdraw a standing ask only while its session remains observed.",
+    schema: {
+      description: "Remove a notification request for an observed session.",
+      parameters: {
+        type: "object",
+        properties: { ...SESSION_IDENTITY_PARAMETERS },
+        required: ["provider_id", "provider_session_id"],
+      },
+    },
+    validate: validateWithdrawSessionNotice,
+  },
+  READ_SESSION_TRANSCRIPT: {
+    name: "read_session_transcript",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.READ_TRANSCRIPT,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(
+      SESSION_TOOL_KIND.READ_TRANSCRIPT,
+      (action, sessions) =>
+        `read ${observedSessionName(action.identity, sessions)}'s transcript aloud`,
+    ),
+    guide: "Read only a local transcript this build documents rendering.",
+    schema: {
+      description: "Read the recent transcript of an observed local session.",
+      parameters: {
+        type: "object",
+        properties: { ...SESSION_IDENTITY_PARAMETERS },
+        required: ["provider_id", "provider_session_id"],
+      },
+    },
+    validate: validateReadSessionTranscript,
+  },
+  CREATE_WORKSPACE: {
+    name: "create_workspace",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.CREATE_WORKSPACE,
+    validatedAgainst: ACT_VALIDATION_TARGET.WORKSPACE_PROJECT,
+    narration: narrate(
+      SESSION_TOOL_KIND.CREATE_WORKSPACE,
+      (action) => `asked ${action.providerId} to create a workspace`,
+    ),
+    guide: "Create only in a project and target from the latest workspace roster.",
+    schema: {
+      description: "Create a workspace for a new agent.",
+      parameters: {
+        type: "object",
+        properties: {
+          provider_id: {
+            type: "string",
+            description: "The provider ID; omit it to create in the default provider.",
+          },
+          project_id: {
+            type: "string",
+            description: "The project ID; omit it to create in that provider's default project.",
+          },
+          target_id: {
+            type: "string",
+            description: "The target ID.",
+          },
+          agent: {
+            type: "string",
+            description: "The agent kind.",
+          },
+          name: {
+            type: "string",
+            description: "An optional workspace name.",
+          },
+          task: {
+            type: "string",
+            description: "An optional opening task.",
+          },
+          model: {
+            type: "string",
+            description: "An optional model.",
+          },
+          effort: {
+            type: "string",
+            description: "An optional effort level.",
+          },
+        },
+        required: [],
+      },
+    },
+    validate: validateCreateWorkspace,
+  },
+  ADD_WORKSPACE_AGENT: {
+    name: "add_workspace_agent",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.ADD_AGENT,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(
+      SESSION_TOOL_KIND.ADD_AGENT,
+      (action, sessions) =>
+        `added a ${action.agent} agent to ${observedSessionName(action.identity, sessions)}`,
+    ),
+    guide: "Add only an agent kind and model the latest session row advertises.",
+    schema: {
+      description: "Add an agent to an observed workspace.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          agent: {
+            type: "string",
+            description: "The agent kind.",
+          },
+          name: {
+            type: "string",
+            description: "An optional agent name.",
+          },
+          task: {
+            type: "string",
+            description: "An optional opening task.",
+          },
+          model: {
+            type: "string",
+            description: "An optional model.",
+          },
+          effort: {
+            type: "string",
+            description: "An optional effort level.",
+          },
+        },
+        required: ["provider_id", "provider_session_id", "agent"],
+      },
+    },
+    validate: validateAddWorkspaceAgent,
+  },
+  RENAME_WORKSPACE: {
+    name: "rename_workspace",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.RENAME_WORKSPACE,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(
+      SESSION_TOOL_KIND.RENAME_WORKSPACE,
+      (action, sessions) =>
+        `renamed the workspace of ${observedSessionName(action.identity, sessions)} to "${action.name}"`,
+    ),
+    guide: "Rename only the workspace target the latest session row advertises.",
+    schema: {
+      description:
+        "Rename the workspace one observed session runs in, to a name the developer just " +
+        "chose — their own words, never a name composed for them. Only sessions whose roster " +
+        "entry says the workspace can be renamed take one.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          name: {
+            type: "string",
+            description: "The workspace's new name, exactly as the developer chose it.",
+          },
+        },
+        required: ["provider_id", "provider_session_id", "name"],
+      },
+    },
+    validate: validateRenameWorkspace,
+  },
+  RENAME_SESSION: {
+    name: "rename_session",
+    family: REALTIME_TOOL_FAMILY.SESSION,
+    actionKind: SESSION_TOOL_KIND.RENAME_SESSION,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(
+      SESSION_TOOL_KIND.RENAME_SESSION,
+      (action, sessions) =>
+        `renamed ${observedSessionName(action.identity, sessions)} to "${action.name}"`,
+    ),
+    guide: "Rename only a chat whose latest roster row advertises rename.",
+    schema: {
+      description:
+        "Rename one observed chat itself — not the workspace around it — to a name the " +
+        "developer just chose, in their own words. Only chats whose roster entry says they can " +
+        "be renamed take one; an ask that names the workspace renames the workspace instead.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...SESSION_IDENTITY_PARAMETERS,
+          name: {
+            type: "string",
+            description: "The chat's new name, exactly as the developer chose it.",
+          },
+        },
+        required: ["provider_id", "provider_session_id", "name"],
+      },
+    },
+    validate: validateRenameSession,
+  },
+  UPDATE_ISSUE_STATE: {
+    name: "update_issue_state",
+    family: REALTIME_TOOL_FAMILY.ISSUE,
+    actionKind: ISSUE_TOOL_KIND.ISSUE_STATE,
+    validatedAgainst: ACT_VALIDATION_TARGET.ISSUE_ROSTER,
+    narration: narrate(
+      ISSUE_TOOL_KIND.ISSUE_STATE,
+      (action) => `moved issue ${action.identity.identifier} to "${action.transition.name}"`,
+    ),
+    guide: "Move an issue only through a transition its latest tracker row advertises.",
+    schema: {
+      description: "Update a tracked issue's state.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...ISSUE_IDENTITY_PARAMETERS,
+          state: {
+            type: "string",
+            description: "The target state.",
+          },
+        },
+        required: ["tracker_id", "issue_id", "state"],
+      },
+    },
+    validate: validateUpdateIssueState,
+  },
+  COMMENT_ON_ISSUE: {
+    name: "comment_on_issue",
+    family: REALTIME_TOOL_FAMILY.ISSUE,
+    actionKind: ISSUE_TOOL_KIND.ISSUE_COMMENT,
+    validatedAgainst: ACT_VALIDATION_TARGET.ISSUE_ROSTER,
+    narration: narrate(
+      ISSUE_TOOL_KIND.ISSUE_COMMENT,
+      (action) => `commented on issue ${action.identity.identifier}`,
+    ),
+    guide: "Add the developer's words only to an issue that currently accepts comments.",
+    schema: {
+      description: "Add a comment to a tracked issue.",
+      parameters: {
+        type: "object",
+        properties: {
+          ...ISSUE_IDENTITY_PARAMETERS,
+          body: {
+            type: "string",
+            description: "The comment to add.",
+          },
+        },
+        required: ["tracker_id", "issue_id", "body"],
+      },
+    },
+    validate: validateCommentOnIssue,
+  },
+  CHANGE_APP_SETTING: {
+    name: "change_app_setting",
+    family: REALTIME_TOOL_FAMILY.APP,
+    actionKind: APP_TOOL_KIND.SETTING,
+    validatedAgainst: ACT_VALIDATION_TARGET.SETTING_ID,
+    narration: narrate(APP_TOOL_KIND.SETTING, (action) => `changed ${action.setting}`),
+    guide: "Change only an adjustable setting and value listed by the current guide.",
+    schema: {
+      description: "Change a Luke setting.",
+      parameters: {
+        type: "object",
+        properties: {
+          setting_id: {
+            type: "string",
+            description: "The setting ID.",
+          },
+          value: {
+            type: "string",
+            description: "The new value.",
+          },
+          effort: {
+            type: "string",
+            description: "An optional effort level.",
+          },
+        },
+        required: ["setting_id", "value"],
+      },
+    },
+    validate: validateChangeAppSetting,
+  },
+  SHOW_PANEL: {
+    name: "show_panel",
+    family: REALTIME_TOOL_FAMILY.APP,
+    actionKind: APP_TOOL_KIND.PANEL,
+    validatedAgainst: ACT_VALIDATION_TARGET.SESSION_ROSTER,
+    narration: narrate(APP_TOOL_KIND.PANEL, (action) => `showed the ${action.tab} panel`),
+    guide: "Show a panel tab and only roster-backed session filters.",
+    schema: {
+      description:
+        "Show Luke's panel on a tab — and, on the sessions tab, narrow or reorder the list. " +
+        'An ask to show, see, or list sessions or agents of some kind — "show me the Codex ' +
+        'agents", "show me my local sessions" — is this tool with a filter, not open_session.',
+      parameters: {
+        type: "object",
+        properties: {
+          tab: {
+            type: "string",
+            enum: Object.values(APP_PANEL_TAB),
+            description: "The tab to show. Defaults to sessions.",
+          },
+          filters: {
+            type: "array",
+            items: { type: "string", enum: SESSION_LIST_FILTER_VALUES },
+            description: SESSION_LIST_FILTER_DESCRIPTION,
+          },
+          sort: {
+            type: "string",
+            enum: Object.values(SESSION_LIST_SORT),
+            description:
+              `Reorders the session list: ${SESSION_LIST_SORT.URGENCY} puts what needs the ` +
+              `developer first, ${SESSION_LIST_SORT.RECENCY} puts what moved last first.`,
+          },
+          query: {
+            type: "string",
+            description:
+              "Optional words to search the session list for; only rows saying every word stay.",
+          },
+        },
+        required: [],
+      },
+    },
+    validate: validateShowPanel,
+  },
+  OPEN_FEEDBACK_COMPOSER: {
+    name: "open_feedback_composer",
+    family: REALTIME_TOOL_FAMILY.APP,
+    actionKind: APP_TOOL_KIND.FEEDBACK,
+    validatedAgainst: ACT_VALIDATION_TARGET.APP_GUIDE,
+    narration: narrate(
+      APP_TOOL_KIND.FEEDBACK,
+      (action) => `opened the ${action.composer} composer`,
+    ),
+    guide: "Open the chosen feedback composer without sending anything.",
+    schema: {
+      description: "Open the feedback composer.",
+      parameters: {
+        type: "object",
+        properties: {
+          kind: {
+            type: "string",
+            enum: Object.values(FEEDBACK_COMPOSER_KIND),
+            description: "The feedback type.",
+          },
+          draft: {
+            type: "string",
+            description: "An optional draft.",
+          },
+        },
+        required: ["kind"],
+      },
+    },
+    validate: validateOpenFeedbackComposer,
+  },
+  RUN_UPDATE_ACTION: {
+    name: "run_update_action",
+    family: REALTIME_TOOL_FAMILY.APP,
+    actionKind: APP_TOOL_KIND.UPDATE,
+    validatedAgainst: ACT_VALIDATION_TARGET.UPDATE_ROW,
+    narration: narrate(APP_TOOL_KIND.UPDATE, (action) => `ran the Updates row's ${action.act}`),
+    guide: "Run only the action the current Updates row offers.",
+    schema: {
+      description:
+        "Press the Updates row's button for the developer: check for updates, open the latest " +
+        "release's page in the browser to download by hand, or restart into an update already " +
+        "downloaded. Only the act the button currently offers runs — the app guide's Updates " +
+        "line names it.",
+      parameters: {
+        type: "object",
+        properties: {
+          action: {
+            type: "string",
+            enum: Object.values(APP_UPDATE_ACT),
+            description: "The act to run, as the guide's Updates line offers it.",
+          },
+        },
+        required: ["action"],
+      },
+    },
+    validate: validateRunUpdateAction,
+  },
+} as const satisfies Record<string, RealtimeToolSpec>;
+
+/** The history sentence declared by the same row that declared the act. */
+export function actNarration(action: CarriedAct, sessions: readonly NormalizedSession[]): string {
+  const row = Object.values(ACTS).find((candidate) => candidate.actionKind === action.kind);
+  return row?.narration(action, sessions) ?? "carried an act";
+}
+
+type RealtimeToolProjection<T extends Record<string, RealtimeToolSpec>> = {
+  [K in keyof T]: Pick<T[K], "name" | "family" | "schema" | "validate">;
+};
+
+function realtimeToolsFromActs<T extends Record<string, RealtimeToolSpec>>(
+  acts: T,
+): RealtimeToolProjection<T> {
+  // SAFETY: every projected entry is constructed from exactly the four fields
+  // required by RealtimeToolProjection, while Object.fromEntries preserves the
+  // keys supplied by Object.entries.
+  return Object.fromEntries(
+    Object.entries(acts).map(([key, act]) => [
+      key,
+      { name: act.name, family: act.family, schema: act.schema, validate: act.validate },
+    ]),
+  ) as RealtimeToolProjection<T>;
+}
+
+/** The Realtime API sees the tool-only projection of the complete act registry. */
+export const REALTIME_TOOLS = realtimeToolsFromActs(ACTS);
+
+function namesFromToolTable<T extends Record<string, { readonly name: string }>>(table: T) {
+  // SAFETY: keys are drawn from the same table object; each entry's name field is the tool id.
+  const names = {} as { [K in keyof T]: T[K]["name"] };
+  // SAFETY: Object.keys returns string[]; every key exists on table because keys are table's own keys.
+  for (const key of Object.keys(table) as (keyof T & string)[]) {
+    const tool = table[key];
+    if (tool) names[key] = tool.name;
+  }
+  return names;
+}
+
+export const REALTIME_TOOL = namesFromToolTable(ACTS);
+
+export type RealtimeToolName = (typeof REALTIME_TOOL)[keyof typeof REALTIME_TOOL];
+
+const REALTIME_TOOL_LIST: readonly RealtimeToolSpec[] = Object.values(ACTS);
+
+const ACTS_BY_NAME = new Map<string, RealtimeToolSpec>(
+  REALTIME_TOOL_LIST.map((tool) => [tool.name, tool]),
+);
+
+export function actValidationTarget(name: string): ActValidationTarget | undefined {
+  return ACTS_BY_NAME.get(name)?.validatedAgainst;
+}
+
+function toolNamesOfFamily(family: RealtimeToolFamily): ReadonlySet<string> {
+  return new Set(
+    REALTIME_TOOL_LIST.filter((tool) => tool.family === family).map((tool) => tool.name),
+  );
+}
+
+const SESSION_TOOL_NAMES = toolNamesOfFamily(REALTIME_TOOL_FAMILY.SESSION);
+const ISSUE_TOOL_NAMES = toolNamesOfFamily(REALTIME_TOOL_FAMILY.ISSUE);
+const APP_TOOL_NAMES = toolNamesOfFamily(REALTIME_TOOL_FAMILY.APP);
+
+/** Whether a tool call names one of the session acts. */
+export function isSessionToolName(name: string): boolean {
+  return SESSION_TOOL_NAMES.has(name);
+}
+
+/** Whether a tool call names one of the issue acts. */
+export function isIssueToolName(name: string): boolean {
+  return ISSUE_TOOL_NAMES.has(name);
+}
+
+/** Whether a tool call is about the app itself rather than about a session. */
+export function isAppToolCall(call: RealtimeFunctionCall): boolean {
+  return APP_TOOL_NAMES.has(call.name);
+}
+
+/** The family a named tool belongs to, or nothing when no such tool exists. */
+export function realtimeToolFamily(name: string): RealtimeToolFamily | undefined {
+  return ACTS_BY_NAME.get(name)?.family;
+}
+
+const SPOKEN_CARDINAL = {
+  0: "zero",
+  1: "one",
+  2: "two",
+  3: "three",
+  4: "four",
+  5: "five",
+  6: "six",
+  7: "seven",
+  8: "eight",
+  9: "nine",
+  10: "ten",
+  11: "eleven",
+  12: "twelve",
+  13: "thirteen",
+  14: "fourteen",
+  15: "fifteen",
+  16: "sixteen",
+  17: "seventeen",
+  18: "eighteen",
+  19: "nineteen",
+  20: "twenty",
+} as const;
+
+/** How many tools Luke has, as he says it in the standing instructions. */
+export function spokenRealtimeToolCount(): string {
+  const count = REALTIME_TOOL_LIST.length;
+  // SAFETY: count is a small cardinal; SPOKEN_CARDINAL keys are the documented spoken counts.
+  return SPOKEN_CARDINAL[count as keyof typeof SPOKEN_CARDINAL] ?? String(count);
+}
+
+/** The tool schemas a Realtime session is configured with. */
+export function realtimeToolDefinitions(): readonly RealtimeToolWireDefinition[] {
+  return REALTIME_TOOL_LIST.map((tool) => ({
+    type: "function",
+    name: tool.name,
+    description: tool.schema.description,
+    parameters: tool.schema.parameters,
+  }));
+}
+
+/**
+ * Validates one tool call against the sessions actually being observed. This
+ * is the renderer's half of the gauntlet — the main process re-validates
+ * against its registry — and it exists so a call the model composed can only
+ * name a session Luke was shown, doing something that session advertised.
+ * Everything else is refused with a reason Luke can say aloud.
+ */
+export function sessionToolAction(
+  call: RealtimeFunctionCall,
+  sessions: readonly NormalizedSession[],
+  workspaceProjects: readonly ObservedWorkspaceProject[] = [],
+  // The models a creation ask may name, per provider — the app's own
+  // build-documented tables, handed in so this stays brand-neutral. The
+  // default offers none, so an ask that names a model is refused rather than
+  // forwarded unchecked.
+  agentModels: (providerId: string) => readonly WorkspaceAgentModels[] = () => [],
+  // The developer's saved creation tie-breaks, riding in beside the projects
+  // they narrow — see {@link SessionToolContext}.
+  defaultProviderId?: string,
+  defaultProjectIds?: Readonly<Partial<Record<string, string>>>,
+): SessionToolAction {
+  const parsed = parseToolArguments(call);
+  if (!parsed.ok) return { status: ACT_RESULT_STATUS.REJECTED, reason: parsed.reason };
+  const tool = ACTS_BY_NAME.get(call.name);
+  if (!tool || tool.family !== REALTIME_TOOL_FAMILY.SESSION) {
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: "No such tool exists." };
+  }
+  return tool.validate(parsed.value, {
+    sessions,
+    workspaceProjects,
+    agentModels,
+    defaultProviderId,
+    defaultProjectIds,
+  });
+}
+
+/**
+ * Validates one issue tool call against the issues actually observed. The
+ * renderer's half of the same gauntlet the session tools run — the main
+ * process re-validates against what it observed — so a call the model
+ * composed can only name an issue Luke was shown, going somewhere its
+ * tracker advertised. Everything else is refused with a reason Luke can say
+ * aloud.
+ */
+export function issueToolAction(
+  call: RealtimeFunctionCall,
+  issues: readonly TrackedIssue[],
+): IssueToolAction {
+  const parsed = parseToolArguments(call);
+  if (!parsed.ok) return { status: ACT_RESULT_STATUS.REJECTED, reason: parsed.reason };
+  const tool = ACTS_BY_NAME.get(call.name);
+  if (!tool || tool.family !== REALTIME_TOOL_FAMILY.ISSUE) {
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: "No such tool exists." };
+  }
+  return tool.validate(parsed.value, { issues });
+}
+
+/**
+ * Validates one app tool call against the guide the app actually provided and
+ * the sessions actually observed. The same posture as {@link sessionToolAction}:
+ * a call the model composed can only name a setting the guide lists, changing
+ * it to a value the guide accepts, a panel view the roster can fill, or the
+ * composer on one of its own two kinds — and a setting the guide marks as
+ * by-hand-only is refused with the path to it, so the refusal Luke voices is
+ * itself the guidance.
+ */
+export function appToolAction(
+  call: RealtimeFunctionCall,
+  guide: AppGuideSnapshot,
+  sessions: readonly NormalizedSession[],
+): AppToolAction {
+  const parsed = parseToolArguments(call);
+  if (!parsed.ok) return { status: ACT_RESULT_STATUS.REJECTED, reason: parsed.reason };
+  const tool = ACTS_BY_NAME.get(call.name);
+  if (!tool || tool.family !== REALTIME_TOOL_FAMILY.APP) {
+    return { status: ACT_RESULT_STATUS.REJECTED, reason: "No such tool exists." };
+  }
+  return tool.validate(parsed.value, { guide, sessions });
+}
+
+/**
+ * Picks the handler for a discriminated `kind`. The map is exhaustive over
+ * the union, so a new kind does not compile until its handler is written.
+ */
+export function dispatchByKind<
+  T extends { kind: string },
+  R,
+  M extends { [K in T["kind"]]: (action: Extract<T, { kind: K }>) => R },
+>(action: T, handlers: M): R {
+  const kind = action.kind;
+  // SAFETY: action.kind is T["kind"]; M is keyed by every member of that union.
+  const handle = handlers[kind as T["kind"]];
+  // SAFETY: M is keyed by T["kind"]; kind selects the handler that accepts this action shape.
+  return handle(action as Extract<T, { kind: typeof kind }>);
+}
