+++ packages/realtime/src/index.ts
@@ -76,6 +76,7 @@ export {
   typedAskEvents,
 } from "./realtime-protocol.js";
 export {
+  type ActEnvelope,
   APP_TOOL_KIND,
   type AppToolAction,
   appToolAction,
@@ -85,6 +86,9 @@ export {
   dispatchByKind,
   type IssueToolAction,
   isAppToolCall,
+  isCarriedAppAction,
+  isCarriedIssueAction,
+  isCarriedSessionAction,
   isIssueToolName,
   isSessionToolName,
   issueToolAction,
